pyproject.toml
[tool.poetry]
name = "poc-dbt-dremio-airflow"
version = "0.1.0"
description = ""
authors = ["Rafael"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pyspark = "^3.5.3"
minio = "^7.2.10"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
